feat: models and mock added
