+++ src/model/offers.js
@@ -1,0 +1,8 @@
+import { getRandomOffer } from '../mock/offers.js';
+
+
+export default class OffersModel {
+  getOffers() {
+    return getRandomOffer().offers;
+  }
+}
